feat: complete Phase 10.2 macro safety enhancements
- Refactor 1000+ line lib.rs into modular structure:
  - controller.rs: Controller macro implementation
  - http_methods.rs: HTTP method macros (GET, POST, etc.)
  - middleware.rs: Middleware application macros
  - params.rs: Parameter and body validation macros
  - routes.rs: Route registration and resource macros
  - groups.rs: Route grouping macros
  - utils.rs: Shared utility functions

- Add comprehensive test coverage for new safety features:
  - Parameter type mismatch validation
  - Invalid route path detection
  - Missing body type specification
  - Type safety validation for #[param] and #[body]

- Improve maintainability: <200 lines per module file
- Maintain 100% test coverage including edge cases
- All macro errors now include actionable hints

Phase 10.2 Complete ✅
- ✅ Arc<Self> async safety pattern
- ✅ Comprehensive error diagnostics
- ✅ Type safety for param/body macros
- ✅ Modular file structure
- ✅ Advanced test coverage
