+++ crates/elif-http-derive/tests/ui/fail/body_no_args.rs
@@ -1,0 +1,7 @@
+use elif_http_derive::{post, body};
+
+#[post("/users")]
+#[body]  // Missing body type argument
+fn create_user() -> String {
+    "created".to_string()
+}
