+++ crates/elif-http-derive/tests/ui/fail/param_type_mismatch.stderr
@@ -1,0 +1,11 @@
+error: Parameter type mismatch: Parameter 'id' has type 'u32' but expected one of: String, &str, str. Hint: Ensure function parameter type matches #[param] declaration.
+ --> tests/ui/fail/param_type_mismatch.rs:5:1
+  |
+5 | fn get_user(id: u32) -> String {  // Function expects u32, param declares string
+  | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+error[E0601]: `main` function not found in crate `$CRATE`
+ --> tests/ui/fail/param_type_mismatch.rs:7:2
+  |
+7 | }
+  |  ^ consider adding a `main` function to `$DIR/tests/ui/fail/param_type_mismatch.rs`
