+++ crates/elif-http-derive/tests/ui/fail/route_path_invalid_param.rs
@@ -1,0 +1,6 @@
+use elif_http_derive::get;
+
+#[get("/users/{}")]  // Empty parameter name
+fn get_user() -> String {
+    "user".to_string()
+}
