+++ crates/elif-http-derive/tests/ui/fail/route_path_invalid_param.stderr
@@ -1,0 +1,11 @@
+error: Invalid route path '/users/{}': empty parameter name '{}'. Hint: Use format like '/users/{id}' with proper parameter syntax.
+ --> tests/ui/fail/route_path_invalid_param.rs:3:7
+  |
+3 | #[get("/users/{}")]  // Empty parameter name
+  |       ^^^^^^^^^^^
+
+error[E0601]: `main` function not found in crate `$CRATE`
+ --> tests/ui/fail/route_path_invalid_param.rs:6:2
+  |
+6 | }
+  |  ^ consider adding a `main` function to `$DIR/tests/ui/fail/route_path_invalid_param.rs`
